+++ readingbat-core/src/main/kotlin/com/readingbat/pages/InvalidRequestPage.kt
@@ -18,7 +18,7 @@
 package com.readingbat.pages
 
 import com.pambrose.common.util.pathOf
-import com.readingbat.common.Endpoints.STATIC_ROOT
+import com.readingbat.common.StaticAssets
 import com.readingbat.common.TwClasses
 import com.readingbat.pages.PageUtils.backLink
 import com.readingbat.pages.PageUtils.bodyTitle
@@ -50,7 +50,7 @@ internal object InvalidRequestPage {
 
           p {
             img(classes = TwClasses.CENTER) {
-              src = pathOf(STATIC_ROOT, "panic.jpg")
+              src = StaticAssets.urlOf("panic.jpg")
               alt = ""
               width = "500"
               height = "313"